Move WritingState object to new file
